requirements.txt
Jinja2==2.7.1
MarkupSafe==0.18
Pygments==1.6
Sphinx==1.2
biopython==1.63
docutils==0.11
nose==1.3.0
wsgiref==0.1.2
